# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"Joe Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/plots":{"get":{"tags":["plot-controller"],"operationId":"getPlots","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotDto"}}}}}}},"put":{"tags":["plot-controller"],"operationId":"postPlot","parameters":[{"name":"plotId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/PutPlotDto"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetPlotDto"}}}}}},"post":{"tags":["plot-controller"],"operationId":"postPlot_1","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/PostPlotDto"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetPlotDto"}}}}}}},"/owners/{ownerId}":{"get":{"tags":["owner-controller"],"operationId":"getOwnerById","parameters":[{"name":"ownerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetOwnerDto"}}}}}},"put":{"tags":["owner-controller"],"operationId":"putOwner","parameters":[{"name":"ownerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/PutOwnerDto"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetOwnerDto"}}}}}}},"/owners":{"get":{"tags":["owner-controller"],"operationId":"getOwners","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetOwnerDto"}}}}}}},"post":{"tags":["owner-controller"],"operationId":"postOwner","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/PostOwnerDto"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetOwnerDto"}}}}}}},"/verification/plotnumber":{"get":{"tags":["verification-controller"],"operationId":"isPlotNumberUnique","parameters":[{"name":"plotNumber","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/plots/{plotId}":{"get":{"tags":["plot-controller"],"operationId":"getPlotById","parameters":[{"name":"plotId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetPlotDto"}}}}}}},"/plots/{ownerId}/owner":{"get":{"tags":["plot-controller"],"operationId":"getPlotByOwnerId","parameters":[{"name":"ownerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotDto"}}}}}}}},"/plots/withHisOwner":{"get":{"tags":["plot-controller"],"operationId":"getPlotsWithHisOwner","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotWithHisOwnerDto"}}}}}}}},"/plots/types":{"get":{"tags":["plot-controller"],"operationId":"getPlotTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotTypeDto"}}}}}}}},"/plots/states":{"get":{"tags":["plot-controller"],"operationId":"getPlotStates","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotStateDto"}}}}}}}},"/plots/availables":{"get":{"tags":["plot-controller"],"operationId":"getAllPlotsAvailables","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotDto"}}}}}}}},"/plotOwners":{"get":{"tags":["plot-owner-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotOwnerDto"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/owners/taxstatus":{"get":{"tags":["owner-controller"],"operationId":"getTaxStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetTaxStatusDto"}}}}}}}},"/owners/plot/{plotId}":{"get":{"tags":["owner-controller"],"operationId":"getOwnersByPlotId","parameters":[{"name":"plotId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerDto"}}}}}}}},"/owners/ownertypes":{"get":{"tags":["owner-controller"],"operationId":"getOwnerTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetOwnerTypeDto"}}}}}}}},"/owners/ownersandplots":{"get":{"tags":["owner-controller"],"operationId":"getOwnersPlots","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetOwnerAndPlot"}}}}}}}},"/owners/ownersandplots/{ownerId}":{"get":{"tags":["owner-controller"],"operationId":"getOwnerAndPlotById","parameters":[{"name":"ownerId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetOwnerAndPlot"}}}}}}},"/owners/dnitypes":{"get":{"tags":["owner-controller"],"operationId":"getDniTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetDniTypeDto"}}}}}}}},"/owners/count-by-status":{"get":{"tags":["owner-controller"],"operationId":"getOwnersCountByStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}},"/owners/allOwnersWithTheirPlots":{"get":{"tags":["owner-controller"],"operationId":"getallOwnersWithTheirPlots","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetOwnerWithHisPlots"}}}}}}}},"/dashboard/plots-stats":{"get":{"tags":["dashboard-controller"],"operationId":"getGeneralStats","parameters":[{"name":"startDate","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"plotType","in":"query","required":false,"schema":{"type":"string"}},{"name":"plotStatus","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlotsStats"}}}}}}},"/dashboard/plots-by-block":{"get":{"tags":["dashboard-controller"],"operationId":"getPlotsByBlock","parameters":[{"name":"startDate","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlotsByBlock"}}}}}}}},"/dashboard/percentage-by-tax-status":{"get":{"tags":["dashboard-controller"],"operationId":"getOwnerPercentageByTaxStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"number","format":"double"}}}}}}}},"/dashboard/owners-distribution":{"get":{"tags":["dashboard-controller"],"operationId":"getOwnershipDistribution","parameters":[{"name":"startDate","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"plotType","in":"query","required":false,"schema":{"type":"string"}},{"name":"plotStatus","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnersPlotsDistribution"}}}}}}}},"/dashboard/count-by-status-per-month":{"get":{"tags":["dashboard-controller"],"operationId":"getOwnersCountByStatusPerMonth","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}}}}},"/dashboard/construction-progress":{"get":{"tags":["dashboard-controller"],"operationId":"getConstructionProgress","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConstructionProgress"}}}}}}}},"/dashboard/blockStats":{"get":{"tags":["dashboard-controller"],"operationId":"getOwnerCountByRole","parameters":[{"name":"startDate","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlockData"}}}}}}}},"/dashboard/Plot-By-Type-Count":{"get":{"tags":["dashboard-controller"],"operationId":"getPlotByTypeCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlotByPlotTypeCountDTO"}}}}}}}},"/dashboard/Plot-By-State-Count":{"get":{"tags":["dashboard-controller"],"operationId":"getPlotByStateCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlotByPlotStateCountDTO"}}}}}}}},"/owners/{id}/{userIdUpdate}":{"delete":{"tags":["owner-controller"],"operationId":"deleteOwner","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"userIdUpdate","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"PutPlotDto":{"required":["built_area_in_m2","plot_state_id","plot_type_id","total_area_in_m2","userUpdateId"],"type":"object","properties":{"total_area_in_m2":{"type":"number","format":"double"},"built_area_in_m2":{"type":"number","format":"double"},"plot_state_id":{"type":"integer","format":"int32"},"plot_type_id":{"type":"integer","format":"int32"},"userUpdateId":{"type":"integer","format":"int32"},"files":{"type":"array","items":{"type":"string","format":"binary"}}}},"FileDto":{"type":"object","properties":{"name":{"type":"string"},"uuid":{"type":"string"}}},"GetPlotDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"plot_number":{"type":"integer","format":"int32"},"block_number":{"type":"integer","format":"int32"},"total_area_in_m2":{"type":"number","format":"double"},"built_area_in_m2":{"type":"number","format":"double"},"plot_state":{"type":"string"},"plot_type":{"type":"string"},"files":{"type":"array","items":{"$ref":"#/components/schemas/FileDto"}}}},"PutOwnerDto":{"required":["active","dni","dniTypeId","email","lastname","name","ownerTypeId","phoneNumber","taxStatusId","userUpdateId"],"type":"object","properties":{"name":{"maxLength":50,"minLength":3,"type":"string"},"lastname":{"maxLength":50,"minLength":3,"type":"string"},"dniTypeId":{"type":"integer","format":"int32"},"dni":{"maxLength":2147483647,"minLength":8,"pattern":"\\d+","type":"string"},"email":{"type":"string"},"phoneNumber":{"maxLength":20,"minLength":10,"pattern":"\\d+","type":"string"},"dateBirth":{"type":"string","format":"date"},"ownerTypeId":{"type":"integer","format":"int32"},"taxStatusId":{"type":"integer","format":"int32"},"businessName":{"maxLength":50,"minLength":0,"type":"string"},"active":{"type":"boolean"},"userUpdateId":{"type":"integer","format":"int32"},"files":{"type":"array","items":{"type":"string","format":"binary"}},"plotId":{"type":"array","items":{"type":"integer","format":"int32"}}}},"GetOwnerDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"lastname":{"type":"string"},"dni_type":{"type":"string"},"dni":{"type":"string"},"dateBirth":{"type":"string","format":"date"},"create_date":{"type":"string","format":"date"},"ownerType":{"type":"string"},"taxStatus":{"type":"string"},"businessName":{"type":"string"},"active":{"type":"boolean"},"files":{"type":"array","items":{"$ref":"#/components/schemas/FileDto"}}}},"PostPlotDto":{"required":["block_number","built_area_in_m2","plot_number","plot_state_id","plot_type_id","total_area_in_m2","userCreateId"],"type":"object","properties":{"plot_number":{"type":"integer","format":"int32"},"block_number":{"type":"integer","format":"int32"},"total_area_in_m2":{"type":"number","format":"double"},"built_area_in_m2":{"type":"number","format":"double"},"plot_state_id":{"type":"integer","format":"int32"},"plot_type_id":{"type":"integer","format":"int32"},"userCreateId":{"type":"integer","format":"int32"},"files":{"type":"array","items":{"type":"string","format":"binary"}}}},"PostOwnerDto":{"required":["active","dni","dni_type_id","email","lastname","name","ownerTypeId","password","phoneNumber","plotId","roles","taxStatusId","userCreateId","username"],"type":"object","properties":{"name":{"maxLength":50,"minLength":3,"type":"string"},"lastname":{"maxLength":50,"minLength":3,"type":"string"},"dni_type_id":{"type":"integer","format":"int32"},"dni":{"maxLength":2147483647,"minLength":8,"pattern":"\\d+","type":"string"},"dateBirth":{"type":"string","format":"date"},"ownerTypeId":{"type":"integer","format":"int32"},"taxStatusId":{"type":"integer","format":"int32"},"businessName":{"maxLength":50,"minLength":0,"type":"string"},"active":{"type":"boolean"},"username":{"maxLength":30,"minLength":1,"type":"string"},"password":{"maxLength":30,"minLength":6,"type":"string"},"email":{"type":"string"},"phoneNumber":{"maxLength":20,"minLength":10,"pattern":"\\d+","type":"string"},"avatarUrl":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"plotId":{"type":"array","items":{"type":"integer","format":"int32"}},"telegramId":{"type":"integer","format":"int32"},"userCreateId":{"type":"integer","format":"int32"},"files":{"type":"array","items":{"type":"string","format":"binary"}}}},"GetPlotWithHisOwnerDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"plot_number":{"type":"integer","format":"int32"},"block_number":{"type":"integer","format":"int32"},"total_area_in_m2":{"type":"number","format":"double"},"built_area_in_m2":{"type":"number","format":"double"},"plot_state":{"type":"string"},"plot_type":{"type":"string"},"files":{"type":"array","items":{"$ref":"#/components/schemas/FileDto"}},"ownerId":{"type":"integer","format":"int32"}}},"GetPlotTypeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"GetPlotStateDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"GetPlotOwnerDto":{"type":"object","properties":{"plot_id":{"type":"integer","format":"int32"},"owner_id":{"type":"integer","format":"int32"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"GetTaxStatusDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"OwnerDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"lastname":{"type":"string"},"dni_type":{"type":"string"},"dni":{"type":"string"},"dateBirth":{"type":"string","format":"date"},"businessName":{"type":"string"},"active":{"type":"boolean"},"ownerType":{"type":"string"},"taxStatus":{"type":"string"},"files":{"type":"array","items":{"$ref":"#/components/schemas/FileDto"}}}},"GetOwnerTypeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"GetOwnerAndPlot":{"type":"object","properties":{"owner":{"$ref":"#/components/schemas/OwnerDto"},"plot":{"type":"array","items":{"$ref":"#/components/schemas/GetPlotDto"}},"user":{"$ref":"#/components/schemas/GetUserDto"}}},"GetUserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"lastname":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"phone_number":{"type":"string"},"dni_type":{"type":"string"},"dni":{"type":"string"},"active":{"type":"boolean"},"avatar_url":{"type":"string"},"datebirth":{"type":"string","format":"date"},"create_date":{"type":"string","format":"date"},"roles":{"type":"array","items":{"type":"string"}},"plot_id":{"type":"array","items":{"type":"integer","format":"int32"}},"telegram_id":{"type":"integer","format":"int32"}}},"GetDniTypeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"GetOwnerWithHisPlots":{"type":"object","properties":{"owner":{"$ref":"#/components/schemas/OwnerDto"},"plot":{"type":"array","items":{"type":"integer","format":"int32"}}}},"PlotsStats":{"type":"object","properties":{"totalPlots":{"type":"integer","format":"int64"},"availablePlots":{"type":"integer","format":"int64"},"constructionPlots":{"type":"integer","format":"int64"},"occupiedPlots":{"type":"integer","format":"int64"},"totalArea":{"type":"number","format":"double"},"constructedArea":{"type":"number","format":"double"}}},"PlotsByBlock":{"type":"object","properties":{"blockNumber":{"type":"integer","format":"int64"},"occupied":{"type":"integer","format":"int64"},"available":{"type":"integer","format":"int64"},"inConstruction":{"type":"integer","format":"int64"},"total":{"type":"integer","format":"int64"}}},"OwnersPlotsDistribution":{"type":"object","properties":{"ownerName":{"type":"string"},"plotCount":{"type":"integer","format":"int64"},"totalArea":{"type":"number","format":"double"}}},"ConstructionProgress":{"type":"object","properties":{"year":{"type":"integer","format":"int64"},"plotsCount":{"type":"integer","format":"int64"}}},"BlockData":{"type":"object","properties":{"blockNumber":{"type":"integer","format":"int32"},"totalArea":{"type":"number","format":"double"},"builtArea":{"type":"number","format":"double"}}},"PlotByPlotTypeCountDTO":{"type":"object","properties":{"type":{"type":"string"},"count":{"type":"integer","format":"int64"}}},"PlotByPlotStateCountDTO":{"type":"object","properties":{"state":{"type":"string"},"count":{"type":"integer","format":"int64"}}}}}}