# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/get/GetPlotWithHisOwnerDto.java
package ar.edu.utn.frc.tup.lc.iv.dtos.get;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

/**
 * DTO para representar un lote junto con la información de su propietario.
 * Incluye detalles como número de lote, manzana, áreas, estado, tipo de lote
 * y archivos asociados, además del identificador del propietario.
 */
@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class GetPlotWithHisOwnerDto {

    /**
     * Identificador único del lote.
     */
    private int id;

    /**
     * Número de lote.
     */
    private int plot_number;

    /**
     * Número de manzana.
     */
    private int block_number;

    /**
     * Metros cuadrados totales del lote.
     */
    private double total_area_in_m2;

    /**
     * Metros cuadrados construidos del lote.
     */
    private double built_area_in_m2;

    /**
     * Estado del lote (disponible, vendido, etc.).
     */
    private String plot_state;

    /**
     * Tipo de lote (casa, negocio, etc).
     */
    private String plot_type;

    /**
     * Lista de archivos asociados al lote.
     */
    private List<FileDto> files;

    /**
     * Id del propietario dueño del plot.
     */
    private Integer ownerId;
}